Rename datasources.clickhouse.ClickhouseMetaSerializer => datasources.mergetree.MetaSerializer
